requirements.txt
pyyaml
GitPython
bioblend==1.7.0
